Move is_stable_version to avoid circular import and improve tests
